+++ cxgo/actions/assignment.go
@@ -4,10 +4,12 @@ import (
 	. "github.com/skycoin/cx/cx"
 )
 
-func StructLiteralAssignment(to []*CXExpression, from []*CXExpression) []*CXExpression {
+// assignStructLiteralFields converts a struct literal to a series of struct field assignments.
+// For example, `foo = Item{x: 10, y: 20}` is converted to: `foo.x = 10; foo.y = 20;`.
+func assignStructLiteralFields (to []*CXExpression, from []*CXExpression, name string) []*CXExpression {
 	for _, f := range from {
-		f.Outputs[0].Name = to[0].Outputs[0].Name
-
+		f.Outputs[0].Name = name
+		
 		if len(to[0].Outputs[0].Indexes) > 0 {
 			f.Outputs[0].Lengths = to[0].Outputs[0].Lengths
 			f.Outputs[0].Indexes = to[0].Outputs[0].Indexes
@@ -18,6 +20,46 @@ func StructLiteralAssignment(to []*CXExpression, from []*CXExpression) []*CXExpr
 	}
 
 	return from
+}
+
+// StructLiteralAssignment handles struct literals, e.g. `Item{x: 10, y: 20}`, and references to
+// struct literals, e.g. `&Item{x: 10, y: 20}` in assignment expressions.
+func StructLiteralAssignment (to []*CXExpression, from []*CXExpression) []*CXExpression {
+	lastFrom := from[len(from)-1]
+	// If the last expression in `from` is declared as pointer
+	// then it means the whole struct literal needs to be passed by reference.
+	if !hasDeclSpec(GetAssignmentElement(lastFrom.Outputs[0]), DECL_POINTER) {
+		return assignStructLiteralFields(to, from, to[0].Outputs[0].Name)
+	} else {
+		// And we also need an auxiliary variable to point to,
+		// otherwise we'd be trying to assign the fields to a nil value.
+		fOut := lastFrom.Outputs[0]
+		auxName := MakeGenSym(LOCAL_PREFIX)
+		aux := MakeArgument(auxName, lastFrom.FileName, lastFrom.FileLine).AddType(TypeNames[fOut.Type])
+		aux.DeclarationSpecifiers = append(aux.DeclarationSpecifiers, DECL_POINTER)
+		aux.CustomType = fOut.CustomType
+		aux.Size = fOut.Size
+		aux.TotalSize = fOut.TotalSize
+		aux.PreviouslyDeclared = true
+		aux.Package = lastFrom.Package
+
+		declExpr := MakeExpression(nil, lastFrom.FileName, lastFrom.FileLine)
+		declExpr.Package = lastFrom.Package
+		declExpr.AddOutput(aux)
+
+		from = assignStructLiteralFields(to, from, auxName)
+
+		assignExpr := MakeExpression(Natives[OP_IDENTITY], lastFrom.FileName, lastFrom.FileLine)
+		assignExpr.Package = lastFrom.Package
+		out := MakeArgument(to[0].Outputs[0].Name, lastFrom.FileName, lastFrom.FileLine)
+		out.PassBy = PASSBY_REFERENCE
+		out.Package = lastFrom.Package
+		assignExpr.AddOutput(out)
+		assignExpr.AddInput(aux)
+
+		from = append([]*CXExpression{declExpr}, from...)
+		return append(from, assignExpr)
+	}
 }
 
 func ArrayLiteralAssignment(to []*CXExpression, from []*CXExpression) []*CXExpression {
@@ -47,7 +89,21 @@ func ShortAssignment(expr *CXExpression, to []*CXExpression, from []*CXExpressio
 	return append(from, expr)
 }
 
-func Assignment(to []*CXExpression, assignOp string, from []*CXExpression) []*CXExpression {
+// getOutputType tries to determine what's the argument that holds the type that should be
+// returned by a function call.
+// This function is needed because CX has some standard library functions that return TYPE_UNDEFINED
+// arguments. In these cases, the output type depends on its input arguments' type. In the rest of
+// the cases, we can simply use the function's return type.
+func getOutputType (expr *CXExpression) *CXArgument {
+	if expr.Operator.Outputs[0].Type != TYPE_UNDEFINED {
+		return expr.Operator.Outputs[0]
+	}
+
+	return expr.Inputs[0]
+}
+
+// Assignment handles assignment statements with different operators, like =, :=, +=, *=.
+func Assignment (to []*CXExpression, assignOp string, from []*CXExpression) []*CXExpression {
 	idx := len(from) - 1
 
 	if pkg, err := PRGRM.GetCurrentPackage(); err == nil {
@@ -65,20 +121,21 @@ func Assignment(to []*CXExpression, assignOp string, from []*CXExpression) []*CX
 				// then it's a literal
 				sym = MakeArgument(to[0].Outputs[0].Name, CurrentFile, LineNo).AddType(TypeNames[from[idx].Outputs[0].Type])
 			} else {
-				sym = MakeArgument(to[0].Outputs[0].Name, CurrentFile, LineNo).AddType(TypeNames[from[idx].Inputs[0].Type])
-				// sym = MakeArgument(to[0].Outputs[0].Name, CurrentFile, LineNo).AddType(TypeNames[from[idx].Operator.Outputs[0].Type])
+				outTypeArg := getOutputType(from[idx])
 
+				sym = MakeArgument(to[0].Outputs[0].Name, CurrentFile, LineNo).AddType(TypeNames[outTypeArg.Type])
+				
 				if from[idx].IsArrayLiteral {
 					sym.Size = from[idx].Inputs[0].Size
 					sym.TotalSize = from[idx].Inputs[0].TotalSize
 					sym.Lengths = from[idx].Inputs[0].Lengths
 				}
-				if from[idx].Inputs[0].IsSlice {
+				if outTypeArg.IsSlice {
 					// if from[idx].Operator.Outputs[0].IsSlice {
 					sym.Lengths = append([]int{0}, sym.Lengths...)
 				}
 
-				sym.IsSlice = from[idx].Inputs[0].IsSlice
+				sym.IsSlice = outTypeArg.IsSlice
 				// sym.IsSlice = from[idx].Operator.Outputs[0].IsSlice
 			}
 			sym.Package = pkg