+++ cxgo/actions/expressions.go
@@ -1,9 +1,19 @@
 package actions
 
 import (
+	"fmt"
+	"os"
+	
 	. "github.com/skycoin/cx/cx"
 	"github.com/skycoin/skycoin/src/cipher/encoder"
 )
+
+// ReturnExpressions stores the `Size` of the return arguments represented by `Expressions`.
+// For example: `return foo() + bar()` is a set of 3 expressions and they represent a single return argument
+type ReturnExpressions struct {
+	Size          int
+	Expressions   []*CXExpression
+}
 
 func IterationExpressions(init []*CXExpression, cond []*CXExpression, incr []*CXExpression, statements []*CXExpression) []*CXExpression {
 	jmpFn := Natives[OP_JMP]
@@ -71,7 +81,7 @@ func IterationExpressions(init []*CXExpression, cond []*CXExpression, incr []*CX
 	return exprs
 }
 
-func trueJmpExpressions() []*CXExpression {
+func trueJmpExpressions () []*CXExpression {
 	pkg, err := PRGRM.GetCurrentPackage()
 	if err != nil {
 		panic(err)
@@ -87,7 +97,7 @@ func trueJmpExpressions() []*CXExpression {
 	return []*CXExpression{expr}
 }
 
-func BreakExpressions() []*CXExpression {
+func BreakExpressions () []*CXExpression {
 	exprs := trueJmpExpressions()
 	exprs[0].IsBreak = true
 	return exprs
@@ -182,16 +192,14 @@ func resolveTypeForUnd (expr *CXExpression) int {
 	return -1
 }
 
-func UndefinedTypeOperation(leftExprs []*CXExpression, rightExprs []*CXExpression, operator *CXFunction) (out []*CXExpression) {
+func UndefinedTypeOperation (leftExprs []*CXExpression, rightExprs []*CXExpression, operator *CXFunction) (out []*CXExpression) {
 	pkg, err := PRGRM.GetCurrentPackage()
 	if err != nil {
 		panic(err)
 	}
 
 	if len(leftExprs[len(leftExprs)-1].Outputs) < 1 {
-		// name := MakeArgument(MakeGenSym(LOCAL_PREFIX), CurrentFile, LineNo).AddType(TypeNames[leftExprs[len(leftExprs)-1].Inputs[0].Type])
 		name := MakeArgument(MakeGenSym(LOCAL_PREFIX), CurrentFile, LineNo).AddType(TypeNames[resolveTypeForUnd(leftExprs[len(leftExprs)-1])])
-
 		name.Size = leftExprs[len(leftExprs)-1].Operator.Outputs[0].Size
 		name.TotalSize = leftExprs[len(leftExprs)-1].Operator.Outputs[0].Size
 		name.Type = leftExprs[len(leftExprs)-1].Operator.Outputs[0].Type
@@ -202,7 +210,6 @@ func UndefinedTypeOperation(leftExprs []*CXExpression, rightExprs []*CXExpressio
 	}
 
 	if len(rightExprs[len(rightExprs)-1].Outputs) < 1 {
-		// name := MakeArgument(MakeGenSym(LOCAL_PREFIX), CurrentFile, LineNo).AddType(TypeNames[rightExprs[len(rightExprs)-1].Inputs[0].Type])
 		name := MakeArgument(MakeGenSym(LOCAL_PREFIX), CurrentFile, LineNo).AddType(TypeNames[resolveTypeForUnd(rightExprs[len(rightExprs)-1])])
 
 		name.Size = rightExprs[len(rightExprs)-1].Operator.Outputs[0].Size
@@ -293,7 +300,16 @@ func ShorthandExpression(leftExprs []*CXExpression, rightExprs []*CXExpression, 
 }
 
 func UnaryExpression(op string, prevExprs []*CXExpression) []*CXExpression {
-	exprOut := prevExprs[len(prevExprs)-1].Outputs[0]
+	if len(prevExprs[len(prevExprs)-1].Outputs) == 0 {
+		println(CompilationError(CurrentFile, LineNo), "invalid indirection")
+		// needs to be stopped immediately
+		os.Exit(CX_COMPILATION_ERROR)
+	}
+
+	// Some properties need to be read from the base argument
+	// due to how we calculate dereferences at the moment.
+	baseOut := prevExprs[len(prevExprs)-1].Outputs[0]
+	exprOut := GetAssignmentElement(prevExprs[len(prevExprs)-1].Outputs[0])
 	switch op {
 	case "*":
 		exprOut.DereferenceLevels++
@@ -304,19 +320,126 @@ func UnaryExpression(op string, prevExprs []*CXExpression) []*CXExpression {
 		exprOut.DeclarationSpecifiers = append(exprOut.DeclarationSpecifiers, DECL_DEREF)
 		exprOut.IsReference = false
 	case "&":
-		exprOut.PassBy = PASSBY_REFERENCE
+		baseOut.PassBy = PASSBY_REFERENCE
 		exprOut.DeclarationSpecifiers = append(exprOut.DeclarationSpecifiers, DECL_POINTER)
 	case "!":
 		if pkg, err := PRGRM.GetCurrentPackage(); err == nil {
 			expr := MakeExpression(Natives[OP_BOOL_NOT], CurrentFile, LineNo)
 			expr.Package = pkg
 
-			expr.AddInput(prevExprs[len(prevExprs)-1].Outputs[0])
+			expr.AddInput(exprOut)
 
+			prevExprs[len(prevExprs)-1] = expr
+		} else {
+			panic(err)
+		}
+	case "-":
+		if pkg, err := PRGRM.GetCurrentPackage(); err == nil {
+			expr := MakeExpression(Natives[OP_UND_NEG], CurrentFile, LineNo)
+			expr.Package = pkg
+			expr.AddInput(exprOut)
 			prevExprs[len(prevExprs)-1] = expr
 		} else {
 			panic(err)
 		}
 	}
 	return prevExprs
+}
+
+// AssociateReturnExpressions associates the output of `retExprs` to the
+// `idx`th output parameter of the current function.
+func AssociateReturnExpressions (idx int, retExprs []*CXExpression) []*CXExpression {
+	var pkg *CXPackage
+	var fn *CXFunction
+	var err error
+
+	pkg, err = PRGRM.GetCurrentPackage()
+	if err != nil {
+		panic(err)
+	}
+
+	fn, err = pkg.GetCurrentFunction()
+	if err != nil {
+		panic(err)
+	}
+
+	lastExpr := retExprs[len(retExprs)-1]
+
+	outParam := fn.Outputs[idx]
+
+	out := MakeArgument(outParam.Name, CurrentFile, LineNo)
+	out.PreviouslyDeclared = true
+
+	if lastExpr.Operator == nil {
+		lastExpr.Operator = Natives[OP_IDENTITY]
+		
+		lastExpr.Inputs = lastExpr.Outputs
+		lastExpr.Outputs = nil
+		lastExpr.AddOutput(out)
+
+		return retExprs
+	} else if len(lastExpr.Outputs) > 0 {
+		expr := MakeExpression(Natives[OP_IDENTITY], CurrentFile, LineNo)
+		expr.AddInput(lastExpr.Outputs[0])
+		expr.AddOutput(out)
+
+		return append(retExprs, expr)
+	} else {
+		lastExpr.AddOutput(out)
+		
+		return retExprs
+	}
+}
+
+// AddJmpToReturnExpressions adds an jump expression that makes a function stop its execution
+func AddJmpToReturnExpressions (exprs ReturnExpressions) []*CXExpression {
+	var pkg *CXPackage
+	var fn *CXFunction
+	var err error
+
+	pkg, err = PRGRM.GetCurrentPackage()
+	if err != nil {
+		panic(err)
+	}
+
+	fn, err = pkg.GetCurrentFunction()
+	if err != nil {
+		panic(err)
+	}
+
+	retExprs := exprs.Expressions
+	
+	if len(fn.Outputs) != exprs.Size && exprs.Expressions != nil {
+		lastExpr := retExprs[len(retExprs)-1]
+		
+		var plural1 string
+		var plural2 string = "s"
+		var plural3 string = "were"
+		if len(fn.Outputs) > 1 {
+			plural1 = "s"
+		}
+		if exprs.Size == 1 {
+			plural2 = ""
+			plural3 = "was"
+		}
+		
+		println(CompilationError(lastExpr.FileName, lastExpr.FileLine), fmt.Sprintf("function '%s' expects to return %d argument%s, but %d output argument%s %s provided", fn.Name, len(fn.Outputs), plural1, exprs.Size, plural2, plural3))
+	}
+	
+	// expression to jump to the end of the embedding function
+	expr := MakeExpression(Natives[OP_JMP], CurrentFile, LineNo)
+
+	// simulating a label so it gets executed without evaluating a predicate
+	expr.Label = MakeGenSym(LABEL_PREFIX)
+	expr.ThenLines = MAX_INT32
+	expr.Package = pkg
+
+	arg := MakeArgument("", CurrentFile, LineNo).AddType("bool")
+	arg.Package = pkg
+
+	expr.AddInput(arg)
+
+	retExprs = append(retExprs, expr)
+
+	return retExprs
 }