+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
-### v0.6 (NOT YET RELEASED)
+### v0.6.1 (NOT YET RELEASED)
+* Additions
+* Libraries
+* Fixed issues
+  * #54: No compilation error when defining a struct with duplicate fields.
+* Documentation
+* IDE (WiP)
+* Miscellaneous
+
+### v0.6
 * Additions
   * Serialization and deserialization
   * Functions as first-class objects
@@ -23,7 +32,29 @@
   * #40: Slice keeps growing though it's cleared inside the loop
   * #41: Scope not working in loops
   * #50: No compilation error when using an invalid identifier
+  * #51: Silent name clash between packages
+  * #52: Some implicit casts were not being caught at compile time
+  * #53: CX was not catching an error involving invalid indirections
+  * #55: Single character declarations are now allowed
+  * #59: Fields of a struct returned by a function call can now be accessed
+  * #61: No compilation error when passing *i32 as an i32 arg and conversely
+  * #62: No compilation error when dereferencing an i32 var
+  * #63: Fixed a problem where inline initializations didn't work with dereferences
+  * #65: Return statements now work in CX, with and without return arguments
+  * #77: Fixed errors related to sending references of structs to functions and
+        assigning references to struct literals
+  * #101: Using different types in shorthands now throws an error
+  * #104: Dubious error message when indexing an array with a substraction expression
+  * #105: Dubious error message when inline initializing a slice
+  * #108: Solved a bug that occured when two functions were named the same in
+    different packages
   * #131: Problem with struct literals in short variable declarations
+  * #132: Short declarations can now be assigned values coming from function calls
+  * #154: Sending pointers to slices to functions is now possible
+  * #167: Passing the address of a slice element is now possible
+  * #199: Trying to call an undefined function no longer throws a segfault
+  * #214: Fixed an error related to type deduction in references to struct fields
+  * #218: Type checking now works with receiving variables of unexpected types
 * Documentation
   * CONTRIBUTING.md: Information about how to contribute to CX
 * IDE (WiP)